Somewhat less stupid tests, skeleton subclasses
Found a workaround for issue with simply creating tempfile and trying to
initialize NanonisFile from it, since there was still an io buffer
object opened on whatever temporary filename.

Now I create a temp dir and don't delete temp files within and then
cleanup the entire directory in the tearDown() method.

Some skeleton classes for the three subclasses were added as well for a
signal class.